리팩토링: 비즈니스 로직과 뷰를 분리
