+++ src/line-manager.js
@@ -1,12 +1,5 @@
-import {
-    addElement, 
-    addInputElement,
-    addTableElement,
-    addSelectElement, 
-    addClickEventListener, 
-    pageInit,
-    addClickEventInButtons
-} from "./common/elements.js";
+import {setPage, addTableRow} from "./View/line-manager-view.js";
+import {addClickEventListener, addClickEventInButtons} from "./common/elements.js";
 import { 
     isEmpty,
     addItem,
@@ -19,34 +12,19 @@ import words from "./common/words.js";
 
 export default class LineManager {
     constructor() {
-        this.setPage();
-        this.lineTableTbody = document.getElementById(words.LINE_TABLE_TBODY);
+        setPage();
         this.setTableContent();
         addClickEventListener(document.getElementById(words.LINE_ADD_BUTTON), () => {this.addLine()});
     }
-
-    setPage() {
-        pageInit();
-        addElement("h4", words.LINE_NAME, null, null, null);
-        addInputElement(words.LINE_NAME_INPUT, words.LINE_INPUT_ALERT, null);
-        addElement("h4", words.LINE_START_STATION, "class", words.LINE_STATION_TEXT, null);
-        addSelectElement(words.LINE_START_STATION_SELECTOR, null);
-        addElement("br", null, null, null, null);
-        addElement("h4", words.LINE_END_STATION, "class", words.LINE_STATION_TEXT, null);
-        addSelectElement(words.LINE_END_STATION_SELECTOR, null);
-        addElement("p", null, null, null, null);
-        addElement("button", words.LINE_ADD, "id",words.LINE_ADD_BUTTON, null);
-        addElement("h2", words.LINE_LIST, null, null, null);
-        addTableElement([words.LINE_NAME, `${words.LINE_START_STATION}역`, `${words.LINE_END_STATION}역`, words.SETTING], words.LINE_TABLE_TBODY, null);
-    }
    
     deleteLine(deleteButton) {
+        const lineTableTbody = document.getElementById(words.LINE_TABLE_TBODY);
         const deleteRow = deleteButton.parentElement.parentElement;
         const lineName = deleteRow.dataset.lineName;
-        this.lineTableTbody.removeChild(deleteRow);
+        lineTableTbody.removeChild(deleteRow);
         deleteKey(lineName);
         deleteItem(words.LINES, lineName);
-        if(this.lineTableTbody.childElementCount === 0) {
+        if(lineTableTbody.childElementCount === 0) {
             deleteKey(words.LINES);
         }
     }
@@ -58,25 +36,12 @@ export default class LineManager {
         }
     }
 
-    addTableRow(lineName, lineStartStationName, lineEndStationName) {
-        const row = this.lineTableTbody.insertRow();
-        row.setAttribute("data-line-name", lineName);
-        const cell1 = row.insertCell(0);
-        const cell2 = row.insertCell(1);
-        const cell3 = row.insertCell(2);
-        const cell4 = row.insertCell(3);
-        cell1.innerHTML = lineName;
-        cell2.innerHTML = lineStartStationName;
-        cell3.innerHTML = lineEndStationName;
-        addElement("button", words.DELETE, "class", words.LINE_DELETE_BUTTON, cell4);
-    }
-
     setTableContent() {
         const lineList = getItemList(words.LINES);
         if(lineList.length > 0) {
             lineList.forEach(line => {
                 const lineSectionList = getItemList(line);
-                this.addTableRow(line, lineSectionList[0], lineSectionList[lineSectionList.length - 1]);
+                addTableRow(line, lineSectionList[0], lineSectionList[lineSectionList.length - 1]);
             });
             addClickEventInButtons(words.LINE_DELETE_BUTTON, this.confirmDeleteLine.bind(this), false);
         }
@@ -88,7 +53,7 @@ export default class LineManager {
     } 
 
     addLineInTable(lineInputName, lineStartStationName, lineEndStationName) {
-        this.addTableRow(lineInputName, lineStartStationName, lineEndStationName);
+        addTableRow(lineInputName, lineStartStationName, lineEndStationName);
         this.addLineSection(lineInputName, lineStartStationName, lineEndStationName);
         addClickEventInButtons(words.LINE_DELETE_BUTTON, this.confirmDeleteLine.bind(this), true);
     }
