+++ src/section-manager.js
@@ -1,12 +1,5 @@
-import {
-    addElement, 
-    addInputElement,
-    addTableElement,
-    addSelectElement, 
-    addClickEventListener, 
-    pageInit,
-    addClickEventInButtons
-} from "./common/elements.js";
+import {setPage, addTableRow, setSectionLineElement} from "./View/section-manager-view.js";
+import {addElement, addClickEventInButtons} from "./common/elements.js";
 import { 
     isEmpty,
     addItem,
@@ -20,32 +13,17 @@ export default class SectionManager {
     constructor() {
         this.sectionTableTbody;
         this.lineName;
-        this.setPage();
+        setPage();
         this.setSectionLineMenuButton();
      }
 
-    setPage() {
-        pageInit();
-        addElement("h3", words.SECTION_LINE_SELECT_TEXT, null, null, null);
-    }
-
-    addTableRow(section) {
-        const row = this.sectionTableTbody.insertRow();
-        row.setAttribute("data-section-name", section);
-        const cell1 = row.insertCell(0);
-        const cell2 = row.insertCell(1);
-        const cell3 = row.insertCell(2);
-        cell1.innerHTML = this.sectionTableTbody.childElementCount - 1;
-        cell2.innerHTML = section;
-        addElement("button", words.SECTION_DELETE_TEXT, "class", words.SECTION_DELETE_BUTTON, cell3);
-    }
-
     setTableContent() {
         const sectionList = getItemList(this.lineName);
+        this.sectionTableTbody = document.querySelector(`#${words.SECTION_TABLE_TBODY}`);
         this.sectionTableTbody.innerHTML = "";
         if(sectionList.length > 0) {
             sectionList.forEach(section => {
-                this.addTableRow(section);
+                addTableRow(section, this.sectionTableTbody);
             });
             addClickEventInButtons(words.SECTION_DELETE_BUTTON, this.confirmDeleteSection.bind(this), false);
         }
@@ -102,22 +80,6 @@ export default class SectionManager {
         const alertText = this.getAlertText(order);
         alertText === "" ? this.addSectionInLine(sectionStationName, order) : alert(alertText);
     }
-
-    setSectionLineElement(menuButton) {
-        const sectionManagementPart = document.getElementById(words.SECTION_MANAGEMENT_PART);
-        this.lineName = menuButton.innerText;
-        sectionManagementPart.innerHTML = "";
-        addElement("h3", `${this.lineName} ${words.MANAGEMENT}`, null, null, sectionManagementPart);
-        addElement("h4", words.SECTION_ADD_TEXT, "id", words.SECTION_ADD_TEXT_ID, sectionManagementPart);
-        addSelectElement(words.SECTION_STATION_SELECTOR, sectionManagementPart);
-        addInputElement(words.SECTION_ORDER_INPUT, words.ORDER, sectionManagementPart);
-        addElement("button", words.ENROLLMENT, "id", words.SECTION_ADD_BUTTON, sectionManagementPart);
-        addElement("p", null, null, null, sectionManagementPart);
-        addTableElement([words.ORDER, words.NAME, words.SETTING], words.SECTION_TABLE_TBODY, sectionManagementPart);
-        addClickEventListener(document.getElementById(words.SECTION_ADD_BUTTON), () => {this.addSection()});
-        this.sectionTableTbody = document.querySelector(`#${words.SECTION_TABLE_TBODY}`);
-        this.setTableContent();
-    }
    
     setSectionLineMenuButton() {
         const lineList = getItemList(words.LINES);
@@ -125,7 +87,7 @@ export default class SectionManager {
             lineList.forEach(line => {
                 addElement("button", line, "class", words.SECTION_LINE_MENU_BUTTON, null);
             });
-            addClickEventInButtons(words.SECTION_LINE_MENU_BUTTON, this.setSectionLineElement.bind(this), false);
+            addClickEventInButtons(words.SECTION_LINE_MENU_BUTTON, setSectionLineElement.bind(this), false);
         }
         addElement("div", null, "id", words.SECTION_MANAGEMENT_PART, null);
     }
